+++ CHANGELOG.md
@@ -11,6 +11,10 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Ability to specify start and end indices in `ParamDB.commit_history()`
 
+### Changed
+
+- `ParamDict` can be initialized from key word arguments in addition to dictionaries
+
 ## [0.2.0] (Mar 8 2023)
 
 ### Added