+++ CHANGELOG.md
@@ -7,6 +7,8 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.4.0] (March 28 2023)
+
 ### Added
 
 - ParamDB load_classes parameter can be set to False to load parameter data classes as
@@ -58,7 +60,8 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Database class `ParamDB` to store parameters in a SQLite file
 - Ability to retrieve the commit history as `CommitEntry` objects
 
-[unreleased]: https://github.com/PainterQubits/paramdb/compare/v0.3.0...main
+[unreleased]: https://github.com/PainterQubits/paramdb/compare/v0.4.0...main
+[0.4.0]: https://github.com/PainterQubits/paramdb/releases/tag/v0.4.0
 [0.3.0]: https://github.com/PainterQubits/paramdb/releases/tag/v0.3.0
 [0.2.0]: https://github.com/PainterQubits/paramdb/releases/tag/v0.2.0
 [0.1.0]: https://github.com/PainterQubits/paramdb/releases/tag/v0.1.0